Move to non-associative FIT\Message::fields array (large messy commit)
Lots going on here. Ideally would be cleaned up into smaller more purposeful commits. All of the work has to do with making sure fields are stored as a non-associative array in FIT\Messages. Some variables and functions were moved to/from parents/children. Several older methods that are no longer in use have been removed with this commit.
